Put all blender mappings in the same file
