--- src/SignIn/Api/BlenderMaps.cs
@@ -1,12 +1,0 @@
-﻿using Starcounter;
-
-namespace SignIn.Api
-{
-    internal class BlenderMaps
-    {
-        public void Register()
-        {
-            Blender.MapUri("/signin/app-name", "app-name");
-        }
-    }
-}
