+++ src/SignIn/Api/MainHandlers.cs
@@ -272,14 +272,6 @@ namespace SignIn
 
                 return page;
             }, new HandlerOptions { SelfOnly = true });
-
-            Blender.MapUri("/signin/user", "user"); //expandable icon
-            Blender.MapUri("/signin/signinuser", "userform"); //inline form
-            Blender.MapUri("/signin/signinuser?{?}", "userform-return"); //inline form
-            Blender.MapUri("/signin/admin/settings", "settings");
-            Blender.MapUri("/signin/user/authentication/password/{?}", "authentication-password");
-            Blender.MapUri("/signin/user/authentication/settings/{?}", "authentication-settings");
-            Blender.MapUri("/signin/partial/user/image", "userimage-default");  // default user image
         }
 
         protected void ClearAuthCookie()
