+++ src/SignIn/SignIn.csproj
@@ -68,7 +68,7 @@
   <ItemGroup>
     <Compile Include="Api\CommitHooks.cs" />
     <Compile Include="Api\MainHandlers.cs" />
-    <Compile Include="Api\BlenderMaps.cs" />
+    <Compile Include="Api\BlenderMapping.cs" />
     <Compile Include="Helpers\AuthorizationHelper.cs" />
     <Compile Include="Helpers\MailSettingsHelper.cs" />
     <Compile Include="Helpers\SignInJwt.cs" />
